docs: blog post page with a right-gutter meta column and table of contents
