+++ packages/docs/app/global.css
@@ -95,10 +95,10 @@ html .prose {
   }
 }
 
-/* long-form posts: h3 gets the same rule as h2 */
-.blog-prose {
+/* long-form posts: h3 gets the same rule as h2; html prefix outranks the base html .prose rule */
+html .blog-prose {
   & > h3 {
-    @apply border-b pb-2 mt-10;
+    @apply border-b pb-2 mt-14;
     border-color: var(--color-fd-border);
   }
 }
